refactor: transformed Arb functions in kotest variant to vals and moved them closer to their usage
